requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=3.1.1
google-genai>=1.59.0
PyJWT[crypto]>=2.10.1
python-multipart>=0.0.9
scikit-learn>=1.4.2
lightgbm>=4.3.0
pandas>=2.2.2
numpy>=1.26.4
